Replace old nested dict accessor with a simpler, more performant one.
